requirements.txt
nose
Flask-Script
mock